election-management/src/test/java/infrastructure/resources/CandidateResourceTest.java
package infrastructure.resources;

import api.CandidateApi;
import api.dto.in.CreateCandidate;
import api.dto.out.Candidate;
import io.quarkus.test.InjectMock;
import io.quarkus.test.common.http.TestHTTPEndpoint;
import io.quarkus.test.junit.QuarkusTest;
import jakarta.ws.rs.core.MediaType;
import org.instancio.Instancio;
import org.jboss.resteasy.reactive.RestResponse;
import org.junit.jupiter.api.Test;

import java.util.Arrays;

import static io.restassured.RestAssured.given;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

@QuarkusTest
@TestHTTPEndpoint(CandidateResource.class)
class CandidateResourceTest {
    @InjectMock
    CandidateApi candidateApi;

    @Test
    void create() {
        var inputCandidate = Instancio.create(CreateCandidate.class);

        given().contentType(MediaType.APPLICATION_JSON).body(inputCandidate)
                .when().post()
                .then().statusCode(RestResponse.StatusCode.CREATED);

        verify(candidateApi).create(inputCandidate);
        verifyNoMoreInteractions(candidateApi);
    }

    @Test
    void list() {
        var outputCandidateList
                = Instancio.stream(Candidate.class).limit(4).toList();

        when(candidateApi.list()).thenReturn(outputCandidateList);

        var response = given()
                .when().get()
                .then().statusCode(RestResponse.StatusCode.OK).extract().as(Candidate[].class);

        verify(candidateApi).list();
        verifyNoMoreInteractions(candidateApi);

        assertEquals(
                Arrays.stream(response).toList(),
                outputCandidateList
        );
    }

}